allocates pseudo-tty and discards make open output
when `make shell` into IDE's container, the terminal is not fully
functional. this fixes the problem by adding `-t` to the `docker run`
command to allocate pseudo-tty.

on Linux, `xdg-open` hangs the terminal window it was executed from.
this fixes the problem by discarding output (to stdout and stderr) of
`make open`.